Should've fixed the moving around of classes and files
